requirements-dev.txt
pytest
pytest-mock
pandas
openpyxl
python-dotenv